Refactor project structure and update configuration for sample deployment
