+++ sample/client/index.html
@@ -8,8 +8,9 @@
 </head>
 
 <body>
-  <div id="vue-app"></div> <!-- Vueアプリがここにマウントされる -->
-  <script type="module" src="./sample/main.ts"></script> <!-- エントリーポイント -->
+  <div id="vue-sample"></div> <!-- Vueアプリがここにマウントされる -->
+
+  <script type="module" src="../main.ts"></script> <!-- エントリーポイント -->
 </body>
 
 </html>
