+++ tsconfig.json
@@ -12,14 +12,14 @@
     "paths": {
       "@/*": [
         "src/*"
-      ] // `@`エイリアスの定義
+      ] // TODO: 共通のファイルができればパスを変える
     }
   },
   "include": [
-    "src/**/*",
-    "src/**/*.d.ts",
-    "src/**/*.tsx",
-    "src/**/*.vue"
+    "**/*",
+    "**/*.d.ts",
+    "**/*.tsx",
+    "**/*.vue"
   ],
   "exclude": [
     "node_modules",
